requirements.txt
flask
requests
socket
json
random
threading
python-nmap